Renaming connection file
Renaming to match package:

https://github.com/jimbobbennett/CircuitPython_ESP32Connection/releases/tag/0.2.0

This won't be put into the community bundle as it's a bit specialised
